added some c++ goodness

+++ Paint.cxx
@@ -18,6 +18,8 @@ along with Rendera; if not, write to the Free Software
 Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301 USA
 */
 
+#include <cmath>
+
 #include "Paint.h"
 #include "Bitmap.h"
 #include "Map.h"
@@ -54,7 +56,7 @@ namespace
                    const int x2, const int y2,
                    const int edge, const int trans)
   {
-    float d = sqrtf(fdist(x1, y1, x2, y2));
+    float d = std::sqrt(fdist(x1, y1, x2, y2));
     float s = (255 - trans) / (((3 << edge) >> 1) + 1);
 
     if(s < 1)
